cds/scripts/requirements.txt
oracledb>=1.3.1
python-dotenv>=1.0.0
